+++ R/modActiveBinding.R
@@ -1,19 +1,36 @@
 makeModActiveBinding <- function(sim, mod) {
+  env <- if (.isPackage(fullModulePath = mod, sim = sim)) {
+    #browser()
+    asNamespace(.moduleNameNoUnderscore(mod))
+  } else {
+    sim@.xData$.mods[[mod]]
+  }
+  if (exists("aaaaa", envir = .GlobalEnv, inherits = FALSE))
+    browser()
+
   makeActiveBinding(sym = "mod",
                     fun = activeModBindingFunction,
-                    env = sim@.xData$.mods[[mod]])
+                    env = env)
 }
 
+#' @include helpers.R
 makeParActiveBinding <- function(sim, mod) {
+  env <- if (.isPackage(fullModulePath = mod, sim = sim)) {
+     asNamespace(.moduleNameNoUnderscore(mod))
+  } else {
+    sim@.xData$.mods[[mod]]
+  }
   makeActiveBinding(sym = "Par",
                     fun = activeParBindingFunction,
-                    env = sim@.xData$.mods[[mod]])
+                    env = env)
 }
 
 activeModBindingFunction <- function(value) {
   ret <- NULL
   if (missing(value)) {
     simEnv <- try(whereInStack("sim"), silent = TRUE)
+    if (exists("aaaaa", envir = .GlobalEnv, inherits = FALSE))
+      browser()
     if (!is(simEnv, "try-error")) {
       sim <- try(get("sim", simEnv, inherits = FALSE), silent = TRUE)
       if (!is(sim, "try-error")) {